Fixed all charges list for user.

+++ models/src/main/java/ar/edu/itba/paw/models/dtos/ProductAmountDTO.java
@@ -8,5 +8,5 @@ import lombok.Getter;
 @Getter
 public class ProductAmountDTO {
     Product product;
-    Integer amount;
+    Long amount;
 }
